+++ Cargo.toml
@@ -9,6 +9,7 @@ license-file = "LICENSE.md"
 resolver = "2"
 members = [
   "./crates/context",
+  "./crates/context/config",
   "./crates/identity",
   "./crates/meroctl",
   "./crates/network",
