+++ crates/context/config/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "calimero-context-config"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+bs58.workspace = true
+borsh = { workspace = true, features = ["derive"] }
+ed25519-dalek.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+thiserror.workspace = true
+
+[lints]
+workspace = true
