+++ crates/context/config/src/types.rs
@@ -1,15 +1,25 @@
 use core::fmt;
 use std::borrow::Cow;
 
+use borsh::{BorshDeserialize, BorshSerialize};
 use ed25519_dalek::{Signature, Verifier, VerifyingKey};
-use near_sdk::serde::Deserialize;
-use near_sdk::{bs58, near, serde, serde_json};
+use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
 use crate::repr::{self, Repr, ReprBytes};
 
-#[derive(Debug)]
-#[near(serializers = [borsh, json])]
+#[derive(
+    Eq,
+    Ord,
+    Clone,
+    Debug,
+    PartialEq,
+    PartialOrd,
+    Serialize,
+    Deserialize,
+    BorshSerialize,
+    BorshDeserialize,
+)]
 pub struct Application<'a> {
     pub id: Repr<ApplicationId>,
     pub blob: Repr<BlobId>,
@@ -18,8 +28,7 @@ pub struct Application<'a> {
     pub metadata: ApplicationMetadata<'a>,
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct Identity([u8; 32]);
 
 impl ReprBytes for Identity {
@@ -40,8 +49,7 @@ impl ReprBytes for Identity {
     }
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct SignerId(Identity);
 
 impl ReprBytes for SignerId {
@@ -62,8 +70,7 @@ impl ReprBytes for SignerId {
     }
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct ContextId(Identity);
 
 impl ReprBytes for ContextId {
@@ -84,8 +91,7 @@ impl ReprBytes for ContextId {
     }
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct ContextIdentity(Identity);
 
 impl ReprBytes for ContextIdentity {
@@ -106,8 +112,7 @@ impl ReprBytes for ContextIdentity {
     }
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct BlobId(Identity);
 
 impl ReprBytes for BlobId {
@@ -128,8 +133,7 @@ impl ReprBytes for BlobId {
     }
 }
 
-#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh])]
+#[derive(Eq, Ord, Copy, Debug, Clone, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
 pub struct ApplicationId(Identity);
 
 impl ReprBytes for ApplicationId {
@@ -150,8 +154,19 @@ impl ReprBytes for ApplicationId {
     }
 }
 
-#[derive(Eq, Ord, Debug, Default, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh, json])]
+#[derive(
+    Eq,
+    Ord,
+    Debug,
+    Default,
+    Clone,
+    PartialEq,
+    PartialOrd,
+    Serialize,
+    Deserialize,
+    BorshSerialize,
+    BorshDeserialize,
+)]
 pub struct ApplicationSource<'a>(#[serde(borrow)] pub Cow<'a, str>);
 
 impl ApplicationSource<'_> {
@@ -160,8 +175,19 @@ impl ApplicationSource<'_> {
     }
 }
 
-#[derive(Eq, Ord, Debug, Default, Clone, PartialEq, PartialOrd)]
-#[near(serializers = [borsh, json])]
+#[derive(
+    Eq,
+    Ord,
+    Debug,
+    Default,
+    Clone,
+    PartialEq,
+    PartialOrd,
+    Serialize,
+    Deserialize,
+    BorshSerialize,
+    BorshDeserialize,
+)]
 pub struct ApplicationMetadata<'a>(#[serde(borrow)] pub Repr<Cow<'a, [u8]>>);
 
 impl ApplicationMetadata<'_> {
@@ -188,15 +214,21 @@ impl ReprBytes for Signature {
     }
 }
 
-#[derive(Eq, Ord, Copy, Clone, Debug, PartialEq, PartialOrd)]
-#[near(serializers = [json])]
+#[derive(Debug, Error)]
+pub enum VerificationKeyParseError {
+    #[error(transparent)]
+    LengthMismatch(repr::LengthMismatch),
+    #[error("invalid key: {0}")]
+    InvalidVerificationKey(ed25519_dalek::SignatureError),
+}
+
+#[derive(Eq, Ord, Copy, Clone, Debug, PartialEq, PartialOrd, Serialize, Deserialize)]
 pub enum Capability {
     ManageApplication,
     ManageMembers,
 }
 
-#[derive(Eq, Debug, Clone, PartialEq)]
-#[near(serializers = [json])]
+#[derive(Eq, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Signed<T> {
     payload: Repr<Box<[u8]>>,
     signature: Repr<Signature>,
@@ -221,9 +253,10 @@ impl<E: fmt::Display> fmt::Debug for Error<E> {
     }
 }
 
-impl<T: serde::Serialize> Signed<T> {
+impl<T: Serialize> Signed<T> {
     pub fn new(payload: &T, sign: impl FnOnce(&[u8]) -> Signature) -> serde_json::Result<Self> {
         let payload = serde_json::to_vec(&payload)?;
+
         let signature = sign(&payload);
 
         Ok(Self {
@@ -239,7 +272,9 @@ impl<'a, T: Deserialize<'a>> Signed<T> {
         f: impl FnOnce(&T) -> Result<VerifyingKey, E>,
     ) -> Result<T, Error<E>> {
         let parsed = serde_json::from_slice(&self.payload)?;
+
         let key = f(&parsed).map_err(Error::KeyDerivationError)?;
+
         key.verify(&self.payload, &self.signature)
             .map_or(Err(Error::InvalidSignature), |_| Ok(parsed))
     }
